Current User method is now working.
If you look at the code, it does seem a little hacky. Do get access to Meteor methods we had to declare a getMeteorData mixin. I honestly am not sure how this works, but it does. Check the code. We are also now displaying items of text that could one day be shopping list items.